Avoid creating multiple "client services" in the same client
Previously, we would always create two client services instances:
- A DaemonStopClient
- A regular DaemonClient

Since client services can hold locks, this causes problems because
the two client services think they are the same process and this
should not happen.

Additionally, clients created through TAPI were never properly
stopped. This meant that multiple requests through TAPI would
leave behind their client services and leave open services.

This usually would not cause a problem because all client services
are throw away and do not hold much state.
