+++ platforms/core-runtime/client-services/build.gradle.kts
@@ -45,10 +45,8 @@ dependencies {
     api(libs.jsr305)
 
     implementation(projects.baseAsm)
+    implementation(projects.fileTemp)
     implementation(projects.serviceRegistryBuilder)
-    implementation(libs.guava)
-    implementation(libs.asm)
-    implementation(libs.slf4jApi)
     implementation(projects.buildOperations)
     implementation(projects.buildProcessServices)
     implementation(projects.fileOperations)
@@ -58,6 +56,10 @@ dependencies {
     implementation(projects.time)
     implementation(projects.toolchainsJvmShared)
     implementation(projects.io)
+
+    implementation(libs.guava)
+    implementation(libs.asm)
+    implementation(libs.slf4jApi)
 
     testImplementation(testFixtures(projects.core)) {
         because("ConcurrentSpecification")
